requirements.txt
google-genai>=1.0.0
openai>=1.42.0
anthropic>=0.40.0
pydantic>=2.8.0
python-dotenv>=1.0.0
